Represent every wire quantity as a lossless Decimal (#23) (#48)
* fix!: represent every wire quantity as a lossless Decimal

Order sizes and live-leg quantities were u64, which rejects the
fractional quantities cryptocurrencies trade in and that a corporate
action can leave on an equity position. OrderLeg::quantity and the
position multipliers and restricted quantity were Decimal deserialized
through f64, which is the representation money is not stored in for
exactly the reason it should not be routed through either.

Adds a wire helper that accepts both shapes the venue uses, a quoted
decimal and a JSON number, without touching f64 in either direction.
serde_json's arbitrary_precision feature is enabled through rust_decimal,
so a Number still holds the digits the venue sent.

Serialization keeps emitting JSON numbers rather than Decimal's own
quoted strings. These fields go into order payloads, the previous
implementation sent numbers, and whether the venue also accepts strings
is not something to discover in production.

BREAKING CHANGE: size, quantity, remaining_quantity, multiplier and
restricted_quantity change type from u64 to Decimal or keep Decimal with
a different wire contract.

Closes #23

* fix: widen the wire error wording beyond quantities

The helper is used for multipliers, tick sizes and display factors as
well, so naming quantities in the failure was wrong for most of its
callers.

+++ src/types/order.rs
@@ -249,8 +249,12 @@ pub struct LiveOrderRecord {
     pub time_in_force: TimeInForce,
     /// The type of order (e.g., Limit, Market, Stop).
     pub order_type: OrderType,
-    /// The size of the order (quantity of the underlying asset).
-    pub size: u64,
+    /// The size of the order.
+    ///
+    /// `Decimal` rather than an integer: cryptocurrencies trade in fractions,
+    /// and a corporate action can leave an equity position with one too.
+    #[serde(with = "crate::types::wire::decimal")]
+    pub size: Decimal,
     /// The symbol of the underlying asset being traded.
     pub underlying_symbol: Symbol,
     /// The price of the order.  Uses `rust_decimal` for arbitrary precision
@@ -285,9 +289,11 @@ pub struct LiveOrderLeg {
     /// The trading symbol for this leg.
     pub symbol: Symbol,
     /// The total quantity of the order for this leg.
-    pub quantity: u64,
+    #[serde(with = "crate::types::wire::decimal")]
+    pub quantity: Decimal,
     /// The remaining quantity to be filled for this leg.
-    pub remaining_quantity: u64,
+    #[serde(with = "crate::types::wire::decimal")]
+    pub remaining_quantity: Decimal,
     /// The action associated with this leg (e.g., Buy, Sell).
     pub action: Action,
     /// A vector of strings representing fills for this leg.  Further
@@ -336,8 +342,11 @@ pub struct OrderLeg {
     instrument_type: InstrumentType,
     /// The trading symbol for the instrument.
     symbol: Symbol,
-    /// The quantity of the instrument to be traded.  Serialized as a float.
-    #[serde(with = "rust_decimal::serde::float")]
+    /// The quantity of the instrument to be traded.
+    ///
+    /// Read from either wire shape and never through `f64`, so a fractional
+    /// crypto quantity survives the round trip.
+    #[serde(with = "crate::types::wire::decimal")]
     quantity: Decimal,
     /// The action to be taken (e.g., Buy, Sell).
     action: Action,
@@ -397,7 +406,8 @@ pub struct DryRunRecord {
     /// The type of the dry-run order (e.g., Limit, Market).
     pub order_type: OrderType,
     /// The size of the dry-run order.
-    pub size: u64,
+    #[serde(with = "crate::types::wire::decimal")]
+    pub size: Decimal,
     /// The underlying symbol for the dry-run order.
     pub underlying_symbol: Symbol,
     /// The price of the dry-run order.  Uses arbitrary precision deserialization.
